+++ silo-core/test/foundry/Silo/flashloan/Flashloan.i.sol
@@ -137,7 +137,7 @@ contract FlashloanTest is SiloLittleHelper, Test, Gas {
         vm.prank(address(receiver));
         token0.approve(address(silo0), amount + fee);
 
-        (uint256 daoAndDeployerRevenueBefore,,,,,) = silo0.getSiloStorage();
+        (uint256 daoAndDeployerRevenueBefore,,,,) = silo0.getSiloStorage();
 
         bytes memory data = abi.encodeWithSelector(
             IERC3156FlashBorrower.onFlashLoan.selector, address(this), address(token0), amount, fee, _data
@@ -161,7 +161,7 @@ contract FlashloanTest is SiloLittleHelper, Test, Gas {
             500
         );
 
-        (uint256 daoAndDeployerRevenueAfter,,,,,) = silo0.getSiloStorage();
+        (uint256 daoAndDeployerRevenueAfter,,,,) = silo0.getSiloStorage();
         assertEq(daoAndDeployerRevenueAfter, daoAndDeployerRevenueBefore + fee);
     }