+++ silo-core/test/foundry/lib/SiloStdLib/GetTotalCollateralAssetsWithInterest.t.sol
@@ -29,13 +29,10 @@ contract GetTotalAssetsWithInterestTest is Test {
         uint256 daoFee;
         uint256 deployerFee;
 
-        SILO.getSiloStorageMock({
-            _daoAndDeployerRevenue: 0,
-            _interestRateTimestamp: 0,
-            _interestFraction: 0,
-            _protectedAssets: 0,
+        SILO.getCollateralAndDebtTotalsWithInterestFactionStorageMock({
             _collateralAssets: 0,
-            _debtAssets: 0
+            _debtAssets: 0,
+            _interestFraction: 0
         });
 
         INTEREST_RATE_MODEL.getCompoundInterestRateMock(silo, block.timestamp, 0);
@@ -44,35 +41,26 @@ contract GetTotalAssetsWithInterestTest is Test {
         INTEREST_RATE_MODEL.getCompoundInterestRateMock(silo, block.timestamp, 0.01e18);
         assertEq(SiloStdLib.getTotalCollateralAssetsWithInterest(silo, interestRateModel, daoFee, deployerFee), 0);
 
-        SILO.getSiloStorageMock({
-            _daoAndDeployerRevenue: 0,
-            _interestRateTimestamp: 0,
-            _interestFraction: 0,
-            _protectedAssets: 0,
+        SILO.getCollateralAndDebtTotalsWithInterestFactionStorageMock({
             _collateralAssets: 1000e18,
-            _debtAssets: 0
+            _debtAssets: 0,
+            _interestFraction: 0
         });
 
         assertEq(SiloStdLib.getTotalCollateralAssetsWithInterest(silo, interestRateModel, daoFee, deployerFee), 1000e18);
 
-        SILO.getSiloStorageMock({
-            _daoAndDeployerRevenue: 0,
-            _interestRateTimestamp: 0,
-            _interestFraction: 0,
-            _protectedAssets: 0,
+        SILO.getCollateralAndDebtTotalsWithInterestFactionStorageMock({
             _collateralAssets: 1000e18,
-            _debtAssets: 500e18
+            _debtAssets: 500e18,
+            _interestFraction: 0
         });
 
         assertEq(SiloStdLib.getTotalCollateralAssetsWithInterest(silo, interestRateModel, daoFee, deployerFee), 1005e18);
 
-        SILO.getSiloStorageMock({
-            _daoAndDeployerRevenue: 0,
-            _interestRateTimestamp: 0,
-            _interestFraction: 0,
-            _protectedAssets: 0,
+        SILO.getCollateralAndDebtTotalsWithInterestFactionStorageMock({
             _collateralAssets: 1000e18,
-            _debtAssets: 1000e18
+            _debtAssets: 1000e18,
+            _interestFraction: 0
         });
 
         daoFee = 0.01e18;