+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - silo-core: add reentrancy protection for liquidation call
 - silo-vaults: fix max redeem
 
+### Updated
+- silo-core: store fraction of interest and fees
+
 ## [2.0.1] - 2025-03-19
 ### Fixed
 - silo-vaults: [Guardian role in the vault incentives module](https://github.com/silo-finance/silo-contracts-v2/commit/389b0575d01f33d745876f6cc26747c082de860b)